abstracting the MemoEditForm to its own abstracted page which both TableSummary pages extend
